requirements.txt
git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord.py>=2.2.2, <=2.9.0
rapptz-asqlite==1.0.1
Pillow==9.5.0